Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserSevice.java
@@ -2,8 +2,11 @@ package com.educandoweb.course.services;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,24 +17,34 @@ public class UserSevice {
     @Autowired
     private UserRepository repository;
 
-    public List<User> findAll(){
+    public List<User> findAll() {
         return repository.findAll();
     }
 
-    public User findById(Long id){
-       Optional<User> obj = repository.findById(id);
-       return obj.orElseThrow(()-> new ResourceNotFoundException(id));
+    public User findById(Long id) {
+        Optional<User> obj = repository.findById(id);
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
-    public User insert(User obj){
-      return  repository.save(obj);
+    public User insert(User obj) {
+        return repository.save(obj);
     }
 
-    public void delete(Long id){
-         repository.deleteById(id);
+    public void delete(Long id) {
+    if(!repository.existsById(id)){
+        throw new ResourceNotFoundException(id);
+    }
+        try {
+            repository.deleteById(id);
+        } catch (EmptyResultDataAccessException e){
+           throw  new ResourceNotFoundException(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
+
     }
 
-    public User update(Long id, User obj){
+    public User update(Long id, User obj) {
         User entity = repository.getReferenceById(id);
         updateData(entity, obj);
         return repository.save(entity);
